+++ Changelog.md
@@ -4,6 +4,7 @@
 ### Version 3.0
 * Add a new theme, "Dawn".
 * Add Distance Conversions.
+* Add Volume Conversions.
 * Update the instructions in the Quadratic Equation Solver.
 * Update the interface of the Quadratic Equation Solver.
 * Update the interface of the Triangles window.